+++ efka-jpa/src/main/java/org/regeneration/efkajpa/repository/UsersRepository.java
@@ -3,5 +3,5 @@ package org.regeneration.efkajpa.repository;
         import org.regeneration.efkajpa.entity.Users;
         import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UsersRepository extends JpaRepository<Users, Integer> {
+public interface UsersRepository extends JpaRepository<Users, Long> {
 }
